+++ bulletinboard-ads-java/src/main/java/com/sap/bulletinboard/ads/services/ReviewsServiceClient.java
@@ -2,6 +2,6 @@ package com.sap.bulletinboard.ads.services;
 
 public interface ReviewsServiceClient {
 
-	Double getAverageRating(String userEmail);
+	float getAverageRating(String userEmail);
 
 }
